SSIDS GPU: fix bugs found in review (int64 truncation, warp-sync, SM_3X)
Fixes from a bug review of the GPU kernels:

- int64->int index truncation (corrupts large factorizations; pre-existing
  in the CUDA code): nnz in assemble.cu cu_load_nodes_sc (now int64, matching
  cu_load_nodes), and offc/offa in syrk.cu (multisyrk_type::offc is int64 but
  was read into an int; loadDevToSmem_generic offa widened to int64).

- Warp-synchronous races in dtrsv.h dblkSolve/dblkSolve_trans: a volatile
  shared scalar was shared across lanes relying on 32-lane lockstep. Add
  explicit SPRAL_SYNCWARP barriers (new per-backend macro; HIP needs a 64-bit
  lane mask). Correct on AMD wavefronts and NVIDIA Volta+; still needs
  on-device numerical validation.

- SM_3X was defined as (__CUDA_ARCH__ >= 300), which is undefined (=0) under
  hipcc, so AMD silently took a branch by accident. Now selected explicitly
  per backend in syrk.cu and reorder.cu; syrk launch bounds no longer dropped
  on AMD.

- Guard the min/max helper macros with #ifndef to avoid clashing with the
  toolchain's std::min/std::max.

All kernels compile under nvcc (CUDA 12) and hipcc (ROCm 7.2); the full
libspral.so still links for the AMD backend. The inter-block spin-lock
synchronisation (assemble.cu, dtrsv.h) is left as-is: it needs on-device
validation / a cooperative-launch rewrite. See src/gpu/README.hip.md.

+++ src/ssids/gpu/kernels/assemble.cu
@@ -53,7 +53,7 @@ cu_load_nodes(
    nlist += 2*lndata->offn;
    double *const lval = lndata->lcol;
 
-   for (int i = threadIdx.x; i < nnz; i += blockDim.x) {
+   for (int64_t i = threadIdx.x; i < nnz; i += blockDim.x) {
      // Note: nlist is 1-indexed, not 0 indexed, so we have to adjust
      const int r = (nlist[2*i+1] - 1) % lda; // row index
      const int c = (nlist[2*i+1] - 1) / lda; // col index
@@ -79,7 +79,7 @@ cu_load_nodes_sc(
     const double *aval
 ) {
    lndata += blockIdx.x;
-   const int nnz = lndata->nnz;
+   const int64_t nnz = lndata->nnz; // int64: matches struct field and cu_load_nodes
    const int lda = lndata->lda;
    const int ldl = lndata->ldl;
 
@@ -87,7 +87,7 @@ cu_load_nodes_sc(
    double *const lval = lndata->lcol;
    rlist += lndata->offr;
 
-   for (int i = threadIdx.x; i < nnz; i += blockDim.x) {
+   for (int64_t i = threadIdx.x; i < nnz; i += blockDim.x) {
       // Note: nlist and rlist are 1-indexed, not 0 indexed, so we adjust
       const int r = (nlist[2*i+1] - 1) % lda; // row index
       const int c = (nlist[2*i+1] - 1) / lda; // col index
